services/framework-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0
asyncpg==0.29.0
pyyaml==6.0.1
anthropic==0.40.0
redis==5.0.8
apscheduler==3.10.4
httpx==0.27.0
python-multipart==0.0.9
